chore: rename @helia/core to @helia/utils
Utils is probably a better name as it has a strong implication that
it's not to be used on it's own whereas `@helia/core` is too close
to `ipfs-core` which we did want people to use on it's own.
