+++ .release-please.json
@@ -4,7 +4,6 @@
   "packages": {
     "packages/block-brokers": {},
     "packages/car": {},
-    "packages/core": {},
     "packages/dag-cbor": {},
     "packages/dag-json": {},
     "packages/helia": {},
@@ -16,6 +15,7 @@
     "packages/mfs": {},
     "packages/routers": {},
     "packages/strings": {},
-    "packages/unixfs": {}
+    "packages/unixfs": {},
+    "packages/utils": {}
   }
 }
