--- packages/core/src/version.ts
@@ -1,2 +1,0 @@
-export const version = '2.0.0'
-export const name = 'helia'
