+++ packages/dag-cbor/package.json
@@ -141,7 +141,7 @@
   },
   "dependencies": {
     "@helia/interface": "^3.0.1",
-    "@ipld/dag-cbor": "^9.0.6",
+    "@ipld/dag-cbor": "^9.0.7",
     "@libp2p/interfaces": "^3.3.1",
     "multiformats": "^13.0.0",
     "progress-events": "^1.0.0"
