+++ packages/helia/package.json
@@ -58,10 +58,10 @@
     "@chainsafe/libp2p-noise": "^14.0.0",
     "@chainsafe/libp2p-yamux": "^6.0.1",
     "@helia/block-brokers": "^1.0.0",
-    "@helia/core": "^0.0.0",
     "@helia/delegated-routing-v1-http-api-client": "^2.0.2",
     "@helia/interface": "^3.0.1",
     "@helia/routers": "^0.0.0",
+    "@helia/utils": "^0.0.0",
     "@libp2p/autonat": "^1.0.1",
     "@libp2p/bootstrap": "^10.0.2",
     "@libp2p/circuit-relay-v2": "^1.0.2",
@@ -77,7 +77,7 @@
     "@libp2p/tcp": "^9.0.2",
     "@libp2p/upnp-nat": "^1.0.1",
     "@libp2p/webrtc": "^4.0.3",
-    "@libp2p/websockets": "^8.0.2",
+    "@libp2p/websockets": "^8.0.10",
     "@libp2p/webtransport": "^4.0.3",
     "blockstore-core": "^4.3.8",
     "datastore-core": "^9.2.6",
