+++ packages/helia/src/index.ts
@@ -27,7 +27,7 @@ import { HeliaP2P } from './helia-p2p.js'
 import { libp2pDefaults } from './utils/libp2p-defaults.js'
 import { createLibp2p } from './utils/libp2p.js'
 import type { DefaultLibp2pServices } from './utils/libp2p-defaults.js'
-import type { Helia } from '@helia/core'
+import type { Helia } from '@helia/interface'
 import type { BlockBroker } from '@helia/interface/blocks'
 import type { ComponentLogger, Libp2p } from '@libp2p/interface'
 import type { KeychainInit } from '@libp2p/keychain'
