+++ packages/unixfs/package.json
@@ -167,9 +167,9 @@
     "hamt-sharding": "^3.0.2",
     "ipfs-unixfs": "^11.0.0",
     "ipfs-unixfs-exporter": "^13.1.0",
-    "ipfs-unixfs-importer": "^15.1.0",
+    "ipfs-unixfs-importer": "^15.2.3",
     "it-glob": "^2.0.4",
-    "it-last": "^3.0.1",
+    "it-last": "^3.0.4",
     "it-pipe": "^3.0.1",
     "merge-options": "^3.0.4",
     "multiformats": "^13.0.0",
